Move `EventScreen` tests from `test` to `androidTest`.
