+++ app/src/androidTest/java/com/github/swent/echo/compose/event/CreateEventScreenTest.kt
@@ -1,12 +1,12 @@
 package com.github.swent.echo.compose.event
 
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.activity.compose.setContent
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import androidx.lifecycle.SavedStateHandle
-import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.swent.echo.MainActivity
 import com.github.swent.echo.authentication.AuthenticationService
 import com.github.swent.echo.data.model.Association
 import com.github.swent.echo.data.model.Event
@@ -16,31 +16,29 @@ import com.github.swent.echo.data.model.Tag
 import com.github.swent.echo.data.repository.Repository
 import com.github.swent.echo.ui.navigation.NavigationActions
 import com.github.swent.echo.ui.navigation.Routes
-import com.github.swent.echo.viewmodels.event.EventStatus
 import com.github.swent.echo.viewmodels.event.EventViewModel
+import dagger.hilt.android.testing.HiltAndroidRule
+import dagger.hilt.android.testing.HiltAndroidTest
 import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import java.time.ZonedDateTime
-import junit.framework.TestCase.assertTrue
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.TestCoroutineScheduler
-import kotlinx.coroutines.test.setMain
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
-class EditEventScreenTest {
+@HiltAndroidTest
+class CreateEventScreenTest {
 
-    @get:Rule val composeTestRule = createComposeRule()
+    @get:Rule(order = 0) val hiltRule = HiltAndroidRule(this)
+
+    @get:Rule(order = 1) val composeTestRule = createAndroidComposeRule<MainActivity>()
 
     private val mockedRepository = mockk<Repository>(relaxed = true)
     private val mockedAuthenticationService = mockk<AuthenticationService>(relaxed = true)
     private lateinit var eventViewModel: EventViewModel
+    private val savedEventId = SavedStateHandle(mapOf())
     private val TEST_EVENT =
         Event(
             eventId = "testid",
@@ -48,7 +46,7 @@ class EditEventScreenTest {
             organizer = Association("testid", "testname", "testdesc"),
             title = "test title",
             description = "test description",
-            location = Location("test location", 100.0, 100.0),
+            location = Location("test location", 10.0, 10.0),
             startDate = ZonedDateTime.now(),
             endDate = ZonedDateTime.now(),
             tags = setOf(Tag("1", "tag1")),
@@ -56,61 +54,36 @@ class EditEventScreenTest {
             15,
             0
         )
-    private val savedEventId = SavedStateHandle(mapOf(Pair("eventId", TEST_EVENT.eventId)))
-    val scheduler = TestCoroutineScheduler()
-    val mockedNavActions = mockk<NavigationActions>(relaxed = true)
 
     @Before
     fun init() {
-        Dispatchers.setMain(StandardTestDispatcher(scheduler))
         every { mockedAuthenticationService.getCurrentUserID() } returns TEST_EVENT.eventId
-        every { mockedNavActions.navigateTo(any()) } returns Unit
-        every { mockedNavActions.goBack() } returns Unit
         coEvery { mockedRepository.getEvent(TEST_EVENT.eventId) } returns TEST_EVENT
+        hiltRule.inject()
         eventViewModel = EventViewModel(mockedRepository, mockedAuthenticationService, savedEventId)
-        composeTestRule.setContent {
-            EditEventScreen(eventViewModel = eventViewModel, navigationActions = mockedNavActions)
-        }
-        scheduler.runCurrent()
     }
 
     @Test
     fun backButtonTriggerNavigateBack() {
+        val mockedNavActions = mockk<NavigationActions>(relaxed = true)
+        every { mockedNavActions.goBack() } returns Unit
+        composeTestRule.activity.setContent {
+            CreateEventScreen(eventViewModel = eventViewModel, navigationActions = mockedNavActions)
+        }
         composeTestRule.onNodeWithTag("Back-button").performClick()
         verify { mockedNavActions.goBack() }
     }
 
     @Test
     fun saveButtonTriggerNavigateToMap() {
+        val mockedNavActions = mockk<NavigationActions>(relaxed = true)
+        every { mockedNavActions.navigateTo(any()) } returns Unit
+        composeTestRule.activity.setContent {
+            CreateEventScreen(eventViewModel = eventViewModel, navigationActions = mockedNavActions)
+        }
         eventViewModel.setEvent(TEST_EVENT)
         composeTestRule.onNodeWithTag("Save-button").performScrollTo().performClick()
-        scheduler.runCurrent()
         composeTestRule.waitForIdle()
         verify { mockedNavActions.navigateTo(Routes.MAP) }
-    }
-
-    @Test
-    fun eventScreenContainsExistingEvent() {
-        scheduler.runCurrent()
-        assertTrue(eventViewModel.event.value == TEST_EVENT)
-        assertTrue(eventViewModel.status.value == EventStatus.Saved)
-    }
-
-    @Test
-    fun deleteButtonWithConfirmNavigateToMap() {
-        composeTestRule.onNodeWithTag("delete-button").performScrollTo().performClick()
-        composeTestRule.onNodeWithTag("delete-confirm").performClick()
-        scheduler.runCurrent()
-        composeTestRule.waitForIdle()
-        verify { mockedNavActions.navigateTo(Routes.MAP) }
-    }
-
-    @Test
-    fun deleteButtonWithCancelDoesNotNavigateToMap() {
-        composeTestRule.onNodeWithTag("delete-button").performScrollTo().performClick()
-        composeTestRule.onNodeWithTag("delete-cancel").performClick()
-        scheduler.runCurrent()
-        composeTestRule.waitForIdle()
-        composeTestRule.onNodeWithTag("Save-button").assertIsDisplayed()
     }
 }
